requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

chromadb>=0.4.22
numpy>=1.26.0

pypdf>=3.17.0
PyMuPDF>=1.23.0
python-docx>=1.1.0
pillow>=10.2.0
pytesseract>=0.3.10

sentence-transformers>=2.3.0
torch>=2.1.0
transformers>=4.36.0

langchain>=0.1.0
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1

watchdog>=3.0.0

requests>=2.31.0
aiohttp>=3.9.0

python-multipart>=0.0.6

tqdm>=4.66.0

faiss-cpu>=1.7.4